MKBHD_chatbot/gradio/requirements.txt
langchain
openai
weaviate-client
python-dotenv